pyproject.toml
[project]
name = "multi-agent-agno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.6.4",
    "fastapi>=0.115.13",
    "google-genai>=1.22.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
    "yfinance>=0.2.63",
]
